requirements.txt
streamlit>=1.33.0
python-dotenv>=1.0.1
openai>=1.40.0
faiss-cpu>=1.8.0
numpy>=1.26.0
tiktoken>=0.7.0
